secp256k1-sys: Introduce key module
Introduce a `key` module and move the three current keys into it.
Re-export them at the crate root so this is refactor only, no logic
changes.
